Moved firewall handler to global space
Reorganized Caddyfile.j2; it now also properly serves :80/ip when fqdn
is an ip address.

The task to search for nftables is marked as successful and skipped, so
now the module loading task checks if it has been skipped before using
that task's stdout as its input.

Added all the conntrack states to icmp packets so that the firewall no
longer blocks ipv6 connectivity.